+++ src/OpenTelemetry.Extensions.Propagators/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Unreleased
 
+## 1.3.0
+
+Released 2022-Jun-03
+
+Use own EventSource named "OpenTelemetry.Extensions.Propagators"
+for internal logging.
+
+## 1.3.0-rc.2
+
+Released 2022-June-1
+
 * Initial release. This has been ported as is from
 [OpenTelemetry.Api](../OpenTelemetry.Api/README.md) package.