+++ src/OpenTelemetry.Shims.OpenTracing/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.0.0-rc9.4
+
+Released 2022-Jun-03
+
 * Removes .NET Framework 4.6.1. The minimum .NET Framework version supported is
   .NET 4.6.2.
   ([#3190](https://github.com/open-telemetry/opentelemetry-dotnet/issues/3190))